# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MirrorTech POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: List all active locations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
    post:
      operationId: createLocation
      tags: [locations]
      summary: Create a new location
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
  /locations/{id}:
    get:
      operationId: getLocation
      tags: [locations]
      summary: Get a single location
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
...
  /shifts:
    get:
      operationId: listShifts
      tags: [shifts]
      summary: List shifts
      security:
        - bearerAuth: []
      parameters:
        - name: locationId
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shift"
    post:
      operationId: createShift
      tags: [shifts]
      summary: Create / schedule a shift
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
  /shifts/{id}:
    get:
      operationId: getShift
      tags: [shifts]
      summary: Get a single shift
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shift
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
    patch:
      operationId: updateShift
      tags: [shifts]
      summary: Update a shift (clock in/out, close, etc.)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
    delete:
      operationId: deleteShift
      tags: [shifts]
      summary: Delete a shift
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /momo/initiate:
    post:
      operationId: initiateMoMo
      tags: [momo]
      summary: Initiate a MoMo payment request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoMoInput"
      responses:
        "200":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoMoResponse"
  /momo/status/{reference}:
    get:
      operationId: getMoMoStatus
      tags: [momo]
      summary: Check MoMo payment status
      security:
        - bearerAuth: []
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoMoStatus"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Location:
      type: object
      required: [id, name, isActive]
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    LocationInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
...
    ShiftInput:
      type: object
      required: [userId, locationId, startTime]
      properties:
        userId:
          type: string
        locationId:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        openingCash:
          type: string
        notes:
          type: string
    ShiftUpdate:
      type: object
      properties:
        endTime:
          type: string
        status:
          type: string
        openingCash:
          type: string
        closingCash:
          type: string
        notes:
          type: string
    Shift:
      type: object
      required: [id, userId, locationId, startTime, status]
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: ["string", "null"]
        locationId:
          type: string
        locationName:
          type: ["string", "null"]
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
        status:
          type: string
        openingCash:
          type: ["string", "null"]
        closingCash:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    MoMoInput:
      type: object
      required: [phone, network, amount, reference]
      properties:
        phone:
          type: string
        network:
          type: string
          enum: [MTN, Telecel]
        amount:
          type: string
        reference:
          type: string
        description:
          type: string
    MoMoResponse:
      type: object
      required: [success, reference, status]
      properties:
        success:
          type: boolean
        reference:
          type: string
        status:
          type: string
        message:
          type: string
    MoMoStatus:
      type: object
      required: [reference, status]
      properties:
        reference:
          type: string
        status:
          type: string
        amount:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        message:
          type: ["string", "null"]