# lib/api-zod/src/generated/types/moMoInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MirrorTech POS API
 * OpenAPI spec version: 0.1.0
 */
import type { MoMoInputNetwork } from './moMoInputNetwork';

export interface MoMoInput {
  phone: string;
  network: MoMoInputNetwork;
  amount: string;
  reference: string;
  description?: string;
}