# lib/api-zod/src/generated/types/moMoResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MirrorTech POS API
 * OpenAPI spec version: 0.1.0
 */

export interface MoMoResponse {
  success: boolean;
  reference: string;
  status: string;
  message?: string;
}