# lib/api-zod/src/generated/types/moMoStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MirrorTech POS API
 * OpenAPI spec version: 0.1.0
 */

export interface MoMoStatus {
  reference: string;
  status: string;
  /** @nullable */
  amount?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  message?: string | null;
}