# lib/api-zod/src/generated/types/shiftInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MirrorTech POS API
 * OpenAPI spec version: 0.1.0
 */

export interface ShiftInput {
  userId: string;
  locationId: string;
  startTime: string;
  endTime?: string;
  openingCash?: string;
  notes?: string;
}